Two stars ➝︎ Stars
The SYP marker can be described as being composed of two or of three stars, depending on the treatment of the middle white star. So the text can fairly be made number-agnostic, appropriate for the seeming switch of the flag from two to three. No change in the SVG, only in mnemonic.